member-service/api/requirements.txt
fastapi
uvicorn[stardard]
sqlalchemy
pymysql
python-multipart
passlib[bcrypt]
python-jose[cryptography]
bcrypt
PyJWT
prometheus-fastapi-instrumentator
python-dotenv
